Rework Command to take arbitrary docker commands
Previously there was some awkward packaging of the command to run
in docker, which relied on the Command object knowing about which
shell to use.

Now `Command` doesn't need to care about the command that will be
run in the container, that is resolved in the `interpret` module
instead. It can now run an arbitrary command inide the docker
container.

This additional flexibility is motivated by a desire to unify
the `command` module with `dind`, so the same pieces of logic
don't need to be replicated in slightly different ways. This will
make it much easier to keep the main `floki` container and `dind`
containers in sync.

Signed-off-by: Richard Lupton <richard.lupton@gmail.com>