Refactor so that Dind now uses DockerCommandBuilder
Added further function to `DocekrCommandBuilder` so that `Dind` can be implemented
in terms of `Command`. This allows the functionality of `Command` to be used to
program `Dind`.

Launching `dind` was also moved to the high-level interpreting functions, rather
than being buried deep in one of the command functions. This makes
it clearer when dind is launched.

`DockerCommandBuilder`s can now be instantiated as daemons, which produce a handle to
the daemon docker container. When this handle is dropped, the container is cleaned up.

Signed-off-by: Richard Lupton <richard.lupton@gmail.com>