Allow dind image to be specified in configuration
The simple `dind` boolean for `dind` can now be expanded out
to allow an image to be specified.

```yaml
dind: true
```

will enable `dind` with the `docker:stable-dind` image for dind,
while

```yaml
dind:
  image: dind:custom
```

will enable `dind` using the image `dind:custom` for dind.

The default is for dind to be disabled.

Signed-off-by: Richard Lupton <richard.lupton@gmail.com>