Use dind image as specified in configuration
This plumbs through the previous configuration so that the
`dind` functionality actually makes use of it.

Some simplifications were possible to existing functions where
`&str`s could be passed around in preference to `String` (which
I've sort of adopted as a preferred default - users of `&str`s
cloning them as and when they need to).

Signed-off-by: Richard Lupton <richard.lupton@gmail.com>